+++ src/App.js
@@ -11,13 +11,6 @@ import LightSwitch from "./components/LightSwitch";
 //   );
 // };
 const App = (props) => {
-  //Declare a variable for state called changeText
-  //Set it's initial value to true
-  let [changeText, setChangeText] = useState(props.initialState);
-  //create a function called handleChange that toggles the value of changeText
-  const handleChange = () => {
-    return setChangeText(!changeText);
-  };
 
 //   //Declare a variable called on and set its value to true
 //    const on = true;
@@ -30,12 +23,15 @@ const App = (props) => {
 //     // the value of changeText is T or F. If it is T, the box shows white. If it is false, the box shows yellow
 //     //also, if changeText is T, the box shows the word "off", if it is false, the box shows "on"
   
-    <div className={`box ${changeText ? "white" : "yellow"}`}
-    onClick={() => handleChange()}
-    >
-      {changeText ? "OFF" : "ON"}
+  <div>
+    <LightSwitch></LightSwitch>
+    <LightSwitch></LightSwitch>
+    <LightSwitch></LightSwitch>
+    <LightSwitch></LightSwitch>
+    <LightSwitch></LightSwitch>
+    <LightSwitch></LightSwitch>
+  </div>
 
-     </div>
       
   );
 }
